requirements.txt
appdirs==1.4.4
dj-database-url==0.5.0
Django==3.2.7
django-chartjs==2.2.1
django-guardian==2.4.0
django-mathfilters==1.0.0
django-smtp-ssl==1.0
django-stdimage==5.3.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
olefile==0.46
packaging==21.0
Pillow==8.3.2
progressbar2==3.53.3
psycopg2==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2021.1
requests==2.26.0
six==1.16.0
virtualenv==20.8.0

pip~=21.1.2
wheel~=0.36.2
setuptools~=57.0.0